Update alternative spec to be considered working draft
Making it clear from the top level README.md that this represents
the current working prototype for Chromium in particular.

The goal is to help make it clear that we've thought about the details
of how this could work to any community group that we
present this to for adoption.

Also updated a few minor points in the process.